fileuploader/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-celery-results==2.5.1
django-filter==23.3
djangorestframework==3.14.0
flower==2.0.1
humanize==4.8.0
kombu==5.3.2
Markdown==3.5.1
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
